requirements.txt
pytest
python-dotenv
requests
requests-toolbelt
pytest-encode
